+++ sim/src/insn_impl/fsw.rs
@@ -9,6 +9,6 @@ pub fn fsw(insn: Insn, cpu: &mut Cpu, bus: &mut Bus) -> cpu::Result<u64> {
     let imm = Insn::sign_extend((imm12hi << 5 | imm12lo) as u64, 12);
     let address = (cpu.load(rs1) as u64).wrapping_add(imm as u64);
 
-    bus.write(address, &(cpu.load(rs2) as u32).to_le_bytes())?;
+    bus.write(address, &(cpu.load(rs2) as f32 as u32).to_le_bytes())?;
     Ok(cpu.pc + 4)
 }
